refactor/EV-780/Clean arch folders
